fix warning

+++ renderlib/threading.cpp
@@ -17,7 +17,7 @@
 ///
 ///
 void
-parallel_for(size_t nb_elements, std::function<void(int start, int end)> functor, bool use_threads)
+parallel_for(size_t nb_elements, std::function<void(size_t start, size_t end)> functor, bool use_threads)
 {
   // -------
   unsigned nb_threads_hint = std::thread::hardware_concurrency();
